+++ src/test/java/org/jabref/model/entry/KeywordListTest.java
@@ -1,5 +1,7 @@
 package org.jabref.model.entry;
 
+import java.util.Optional;
+
 import org.junit.Before;
 import org.junit.Test;
 
@@ -87,5 +89,27 @@ public class KeywordListTest {
 
         assertEquals(new KeywordList(expectedOne, expectedTwo),
                 KeywordList.parse("Parent1 > Node1 > Child1, Parent2 > Node2 > Child2", ',', '>'));
+    }
+
+    @Test
+    public void parseEmptyOptionalReturnsEmptyList() throws Exception {
+        assertEquals(new KeywordList(), KeywordList.parse(Optional.empty(), ','));
+    }
+
+    @Test
+    public void parseEmptyStringOptionalReturnsEmptyList() throws Exception {
+        assertEquals(new KeywordList(), KeywordList.parse(Optional.of(""), ','));
+    }
+
+    @Test
+    public void parseOneWordOptionalReturnsOneKeyword() throws Exception {
+        assertEquals(new KeywordList("keywordOne"),
+                KeywordList.parse(Optional.of("keywordOne"), ','));
+    }
+
+    @Test
+    public void parseTwoWordOptionalReturnsTwoKeywords() throws Exception {
+        assertEquals(new KeywordList("keywordOne", "keywordTwo"),
+                KeywordList.parse(Optional.of("keywordOne, keywordTwo"), ','));
     }
 }